Formalizing concept of prior and here.
In language, suggesting that we:

 - Allow the user to explicitly access the value of an attribute prior to the current event. This prevents cyclic computational graphs.
 - Allow the user to ask for other entities “here” where “here” refers to the parent patch. This makes the prior suggestion more explicit.

In engine, suggesting we add getFromPriorStep so that the engine can know 1) if an attribute has been modified in this time step or if it still needs to be calculated while 2) getting the prior value if the the update code refers to it. This also allows us to have the user be more explicit in specifying current versus prior value.