Add support for legacy browsers.
WASM is faster but I did some testing on BrowserStack with older browsers and am convinced that using TeaVM’s fallback JS target for browsers not supporting WASM is a good move. It will test to see if the browser supports wasm first and, if not, it will fall back to the JS target.